Reorganize document util fns
Rename folder to be more generic. Use helpers.ts for handlebar
helpers only.

Ticket: AT-7759
