+++ document-generation/utils/utils.ts
@@ -1,0 +1,24 @@
+import { PeriodUnit, IPeriod } from "../../models/document-generation";
+
+export const capitalize = (string: string) => {
+  if (typeof string !== "string") {
+    return "";
+  }
+  return string.charAt(0).toUpperCase() + string.slice(1).toLocaleLowerCase();
+};
+
+export const convertPeriodToMonths = (period: IPeriod): number => {
+  const { periodUnit, periodUnitCount } = period;
+  switch (periodUnit) {
+    case PeriodUnit.YEAR:
+      return periodUnitCount * 12;
+    case PeriodUnit.MONTH:
+      return periodUnitCount;
+    case PeriodUnit.WEEK:
+      return Math.ceil(periodUnitCount / 4.345);
+    case PeriodUnit.DAY:
+      return Math.ceil(periodUnitCount / 30.4167);
+    default:
+      return 12;
+  }
+};
